chore(doctor): split scope orchestration and refresh CLI docs
